refactor: 팀 조회 서비스 분리
