+++ backend/src/main/java/com/woowacourse/levellog/team/application/TeamService.java
@@ -9,28 +9,17 @@ import com.woowacourse.levellog.team.domain.InterviewRole;
 import com.woowacourse.levellog.team.domain.Participant;
 import com.woowacourse.levellog.team.domain.ParticipantRepository;
 import com.woowacourse.levellog.team.domain.Participants;
-import com.woowacourse.levellog.team.domain.SimpleParticipant;
-import com.woowacourse.levellog.team.domain.SimpleParticipants;
 import com.woowacourse.levellog.team.domain.Team;
-import com.woowacourse.levellog.team.domain.TeamQueryRepository;
 import com.woowacourse.levellog.team.domain.TeamRepository;
 import com.woowacourse.levellog.team.domain.TeamStatus;
-import com.woowacourse.levellog.team.dto.AllParticipantDto;
-import com.woowacourse.levellog.team.dto.AllSimpleParticipantDto;
 import com.woowacourse.levellog.team.dto.InterviewRoleDto;
-import com.woowacourse.levellog.team.dto.ParticipantDto;
-import com.woowacourse.levellog.team.dto.TeamDto;
-import com.woowacourse.levellog.team.dto.TeamListDto;
-import com.woowacourse.levellog.team.dto.TeamSimpleDto;
 import com.woowacourse.levellog.team.dto.TeamStatusDto;
 import com.woowacourse.levellog.team.dto.TeamWriteDto;
-import com.woowacourse.levellog.team.dto.WatcherDto;
 import com.woowacourse.levellog.team.exception.HostUnauthorizedException;
 import com.woowacourse.levellog.team.exception.TeamNotFoundException;
 import com.woowacourse.levellog.team.support.TimeStandard;
 import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -43,10 +32,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class TeamService {
 
-    private static final String ALL_STATUS = "all";
-
     private final TeamRepository teamRepository;
-    private final TeamQueryRepository teamQueryRepository;
     private final ParticipantRepository participantRepository;
     private final MemberRepository memberRepository;
     private final TimeStandard timeStandard;
@@ -63,88 +49,6 @@ public class TeamService {
         participantRepository.saveAll(participants.getValues());
 
         return savedTeam.getId();
-    }
-
-    public TeamListDto findAll(final String status, final int page, final int size) {
-        TeamStatus.checkClosed(status);
-
-        final List<AllSimpleParticipantDto> allParticipants = teamQueryRepository.findAllList(size, page * size);
-        final List<TeamSimpleDto> teamSimpleDtos = allParticipants.stream()
-                .filter(it -> filterStatus(status, toTeamStatus(it)))
-                .collect(Collectors.groupingBy(AllSimpleParticipantDto::getId, LinkedHashMap::new, Collectors.toList()))
-                .values()
-                .stream()
-                .map(it -> TeamSimpleDto.of(it, toTeamStatus(it.get(0))))
-                .collect(Collectors.toList());
-
-        return new TeamListDto(teamSimpleDtos);
-    }
-
-    public TeamListDto findAllByMemberId(final Long memberId) {
-        final Member member = getMember(memberId);
-
-        final List<AllSimpleParticipantDto> allParticipants = teamQueryRepository.findMyList(member);
-        final List<TeamSimpleDto> teamSimpleDtos = allParticipants.stream()
-                .collect(Collectors.groupingBy(AllSimpleParticipantDto::getId, LinkedHashMap::new, Collectors.toList()))
-                .values()
-                .stream()
-                .map(it -> TeamSimpleDto.of(it, toTeamStatus(it.get(0))))
-                .collect(Collectors.toList());
-
-        return new TeamListDto(teamSimpleDtos);
-    }
-
-    public TeamDto findByTeamIdAndMemberId(final Long teamId, final Long memberId) {
-        final List<AllParticipantDto> allParticipants = teamQueryRepository.findAllByTeamId(teamId, memberId);
-        if (allParticipants.isEmpty()) {
-            throw new TeamNotFoundException(DebugMessage.init()
-                    .append("teamId", teamId)
-                    .append("membmerId", memberId));
-        }
-
-        final Team team = allParticipants.get(0).getTeam();
-
-        final SimpleParticipants participants = toSimpleParticipants(allParticipants);
-
-        final TeamStatus status = team.status(timeStandard.now());
-        final boolean isParticipant = participants.isContains(memberId);
-        final List<Long> interviewers = participants.toInterviewerIds(memberId, team.getInterviewerNumber());
-        final List<Long> interviewees = participants.toIntervieweeIds(memberId, team.getInterviewerNumber());
-
-        return TeamDto.from(team, participants.toHostId(), status, isParticipant, interviewers, interviewees,
-                toParticipantDto(allParticipants), toWatcherDtos(allParticipants));
-    }
-
-    private boolean filterStatus(final String status, final TeamStatus teamStatus) {
-        if (!status.equals(ALL_STATUS)) {
-            return teamStatus.isSame(status);
-        }
-        return true;
-    }
-
-    private TeamStatus toTeamStatus(final AllSimpleParticipantDto dto) {
-        return TeamStatus.of(dto.isClosed(), dto.getStartAt(), timeStandard.now());
-    }
-
-    private SimpleParticipants toSimpleParticipants(final List<AllParticipantDto> filteredParticipants) {
-        final List<SimpleParticipant> participants = filteredParticipants.stream()
-                .map(AllParticipantDto::toSimpleParticipant)
-                .collect(Collectors.toList());
-        return new SimpleParticipants(participants);
-    }
-
-    private List<ParticipantDto> toParticipantDto(final List<AllParticipantDto> filteredParticipants) {
-        return filteredParticipants.stream()
-                .filter(it -> !it.isWatcher())
-                .map(ParticipantDto::from)
-                .collect(Collectors.toList());
-    }
-
-    private List<WatcherDto> toWatcherDtos(final List<AllParticipantDto> filteredParticipants) {
-        return filteredParticipants.stream()
-                .filter(AllParticipantDto::isWatcher)
-                .map(WatcherDto::from)
-                .collect(Collectors.toList());
     }
 
     public TeamStatusDto findStatus(final Long teamId) {
