+++ backend/src/main/java/com/woowacourse/levellog/team/presentation/TeamController.java
@@ -2,6 +2,7 @@ package com.woowacourse.levellog.team.presentation;
 
 import com.woowacourse.levellog.authentication.support.Authentic;
 import com.woowacourse.levellog.authentication.support.PublicAPI;
+import com.woowacourse.levellog.team.application.TeamQueryService;
 import com.woowacourse.levellog.team.application.TeamService;
 import com.woowacourse.levellog.team.dto.InterviewRoleDto;
 import com.woowacourse.levellog.team.dto.TeamDto;
@@ -28,6 +29,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class TeamController {
 
     private final TeamService teamService;
+    private final TeamQueryService teamQueryService;
 
     @PostMapping
     public ResponseEntity<Void> save(@RequestBody @Valid final TeamWriteDto teamDto,
@@ -41,7 +43,7 @@ public class TeamController {
     public ResponseEntity<TeamListDto> findAll(@RequestParam(defaultValue = "all") final String status,
                                                @RequestParam(defaultValue = "0") final String page,
                                                @RequestParam(defaultValue = "20") final String size) {
-        final TeamListDto response = teamService.findAll(status, Integer.parseInt(page), Integer.parseInt(size));
+        final TeamListDto response = teamQueryService.findAll(status, Integer.parseInt(page), Integer.parseInt(size));
         return ResponseEntity.ok(response);
     }
 
@@ -49,7 +51,7 @@ public class TeamController {
     @PublicAPI
     public ResponseEntity<TeamDto> findById(@PathVariable final Long teamId,
                                             @Authentic final Long memberId) {
-        final TeamDto response = teamService.findByTeamIdAndMemberId(teamId, memberId);
+        final TeamDto response = teamQueryService.findByTeamIdAndMemberId(teamId, memberId);
         return ResponseEntity.ok(response);
     }
